requirements/base.txt
django==1.7.6
django-solo==1.1.0
django-annoying==0.8.1
django-axes==1.3.9
django-user-accounts==1.0.1
django-custom-user==0.5
